Update unit tests.
